+++ src/Cwaff-Guns/Alligator.cs
@@ -175,7 +175,7 @@ public class Alligator : CwaffGun
             uic.GunAmmoCountLabel.AutoHeight = true; // enable multiline text
             uic.GunAmmoCountLabel.ProcessMarkup = true; // enable multicolor text
             string energy = Mathf.RoundToInt(_ENERGY_MULT * this._alligator.energyProduction).ToString();
-            uic.GunAmmoCountLabel.Text = $"[color #00ffff]{energy}[/color][sprite \"charge_ui\"]\n{this._gun.CurrentAmmo}/{this._gun.AdjustedMaxAmmo}";
+            uic.GunAmmoCountLabel.Text = $"[color #00ffff]{energy}[/color][sprite \"charge_ui\"]\n{this._owner.VanillaAmmoDisplay()}";
             return true;
         }
     }