+++ src/Cwaff-Guns/SubtractorBeam.cs
@@ -144,7 +144,7 @@ public class SubtractorBeamAmmoDisplay : CustomAmmoDisplay
         uic.GunAmmoCountLabel.AutoHeight = true; // enable multiline text
         uic.GunAmmoCountLabel.ProcessMarkup = true; // enable multicolor text
         float damage = this._sub._lastHitEnemy ? this._sub._lastHitEnemy.currentHealth : 0f;
-        uic.GunAmmoCountLabel.Text = $"[color #66dd66]{damage}[/color]\n{this._gun.CurrentAmmo}/{this._gun.AdjustedMaxAmmo}";
+        uic.GunAmmoCountLabel.Text = $"[color #66dd66]{damage}[/color]\n{this._owner.VanillaAmmoDisplay()}";
         return true;
     }
 }